move layout component to Layout folder, fix highlight error
