+++ components/layouts/Layout/Layout.tsx
@@ -1,17 +1,17 @@
 import React, { useState, ReactNode } from 'react';
 import styled from 'styled-components';
 
-import { Header } from './Header/Header';
-import { Footer } from './Footer/Footer';
-import { Button } from '../core/Button/Button';
-import { useScrollPosY } from '../hooks/useScrollPosY';
+import { Header } from '../Header/Header';
+import { Footer } from '../Footer/Footer';
+import { Button } from '../../core/Button/Button';
+import { useScrollPosY } from '../../hooks/useScrollPosY';
 
-import { CDN_URL_STATIC_DIRECTORY } from '../../helpers/utils';
+import { CDN_URL_STATIC_DIRECTORY } from '../../../helpers/utils';
 import {
   SCROLL_DELAY,
   SHOWING_SCROLLUP_BUTTON_HEIGHT,
-} from '../../helpers/staticData';
-import { fadeAnimation } from '../../helpers/animations';
+} from '../../../helpers/staticData';
+import { fadeAnimation } from '../../../helpers/animations';
 
 type LayoutProps = {
   children: ReactNode;
