+++ components/pages/ErrorPage/ErrorPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
 
-import { Layout } from '../../layouts/Layout';
+import { Layout } from '../../layouts/Layout/Layout';
 
 import { BlockTitle } from '../../core/BlockTitle/BlockTitle';
 import { Button } from '../../core/Button/Button';
