+++ components/pages/HomePage/HomePage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Link from 'next/link';
 import styled from 'styled-components';
 
-import { Layout } from '../../layouts/Layout';
+import { Layout } from '../../layouts/Layout/Layout';
 
 import { FullLoader } from '../../ui/FullLoader/FullLoader';
