+++ components/pages/SearchPage/SearchPage.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef } from 'react';
 import Link from 'next/link';
 import styled from 'styled-components';
 
-import { Layout } from '../../layouts/Layout';
+import { Layout } from '../../layouts/Layout/Layout';
 
 import { Filter } from '../../ui/Filter/Filter';
 import { FullLoader } from '../../ui/FullLoader/FullLoader';
